Move rustup-specific code out of rustup-utils::download
